Домашнее задание ко второму занятию - Исправление замечаний(тесты)
