+++ student-test-app-annotations/src/test/java/QuestionParserServiceTest.java
@@ -1,5 +1,5 @@
 import org.junit.Test;
-import ru.otus.studenttestappannotations.domain.Answer;
+import org.mockito.Mockito;
 import ru.otus.studenttestappannotations.domain.Question;
 import ru.otus.studenttestappannotations.service.LineParserServiceCSVImpl;
 import ru.otus.studenttestappannotations.service.QuestionParserService;
@@ -8,9 +8,8 @@ import ru.otus.studenttestappannotations.service.QuestionParserServiceCSVImpl;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertTrue;
 
 public class QuestionParserServiceTest {
     @Test
@@ -21,17 +20,10 @@ public class QuestionParserServiceTest {
             questions.addAll(service.parse(scanner));
         }
         assertTrue(questions.size() == 1);
-        Question question = questions.get(0);
-        assertNotNull(question);
-        assertEquals("Kerry … swim very well yet.", question.getQuestionText());
-        List<Answer> answers = question.getAnswerVariants();
-        assertNotNull(answers);
-        assertTrue(answers.size() == 3);
-        assertArrayEquals(new String[]{"shouldn't", "ought", "cannot"}, answers.stream().map(a -> a.getText()).collect(Collectors.toList()).toArray(new String[3]));
     }
     @Test
     public void parseEmpty() {
-        QuestionParserService service = new QuestionParserServiceCSVImpl(new LineParserServiceCSVImpl());
+        QuestionParserService service = new QuestionParserServiceCSVImpl(Mockito.mock(LineParserServiceCSVImpl.class));
         List<Question> questions = new ArrayList<>();
         try (Scanner scanner = new Scanner("")) {
             questions.addAll(service.parse(scanner));
@@ -40,23 +32,9 @@ public class QuestionParserServiceTest {
     }
     @Test
     public void parseNull() {
-        QuestionParserService service = new QuestionParserServiceCSVImpl(new LineParserServiceCSVImpl());
+        QuestionParserService service = new QuestionParserServiceCSVImpl(Mockito.mock(LineParserServiceCSVImpl.class));
         List<Question> questions = new ArrayList<>();
         questions.addAll(service.parse(null));
         assertTrue(questions.size() == 0);
-    }
-    @Test
-    public void parseFreeQuestion() {
-        QuestionParserService service = new QuestionParserServiceCSVImpl(new LineParserServiceCSVImpl());
-        List<Question> questions = new ArrayList<>();
-        try (Scanner scanner = new Scanner("Kerry … swim very well yet.")) {
-            questions.addAll(service.parse(scanner));
-        }
-        assertTrue(questions.size() == 1);
-        Question question = questions.get(0);
-        assertNotNull(question);
-        assertEquals("Kerry … swim very well yet.", question.getQuestionText());
-        List<Answer> answers = question.getAnswerVariants();
-        assertTrue(answers == null || answers.size() == 0);
     }
 }
